Export the interfaces to a separate file
