+++ src/app/components/common/button/minimize/index.less
