+++ src/app/components/common/iframe/iframe.interface.ts
@@ -1,0 +1,4 @@
+export interface IIframeProps {
+  src: string;
+  isResponsive?: boolean;
+}
