+++ src/app/components/common/list/list-item/teach-me-list-item/TMListItemThumbnail.tsx
@@ -1,14 +1,13 @@
 import React from "react";
 
+// interfaces
+import { ITMListItemThumbnailProps } from "../../list.interface";
+
 export default function TMListItemThumbnail({
   title,
   thumbSrcRatio_1_1,
   thumbSrcRatio_2_1,
-}: {
-  title: string;
-  thumbSrcRatio_1_1: string;
-  thumbSrcRatio_2_1: string;
-}) {
+}: ITMListItemThumbnailProps) {
   return (
     <picture className="thumb">
       <img
