+++ src/app/components/common/message-container/messageContainer.interface.ts
@@ -1,0 +1,7 @@
+import { CourseState } from "../../layout/screens/courses/courses.interface";
+
+export interface IMessageContainerProps {
+  message: string;
+  type: CourseState;
+  className: string;
+}
