+++ src/app/components/common/progress-bar/progressBar.interface.ts
@@ -1,0 +1,6 @@
+export interface IProgressBarProps {
+  percentCompletion: number;
+  showTitle?: boolean;
+  customTitle?: string;
+  showPercentages?: boolean;
+}
