+++ src/app/components/common/user-details/userDetails.interface.ts
@@ -1,0 +1,4 @@
+export interface IUserDetailsProps {
+  greeting?: boolean;
+  progressBar?: boolean;
+}
