+++ src/app/components/layout/screens/courses/courses.interface.ts
