+++ src/app/components/layout/screens/information-screen/informationScreen.interface.ts
